Add Spring-Boot tests: createGame, createPlayer, joinGame
