+++ server/game-server/build.gradle
@@ -13,14 +13,13 @@ version '0.0.1-SNAPSHOT'
 apply plugin: 'war'
 
 war {
-    baseName = 'game-server'
+    baseName = 'game-serv'
     version = '0.0.1'
 }
 
 tasks.named('check') {
     dependsOn tasks.named('testAggregateTestReport', TestReport)
 }
-
 
 configurations {
     providedRuntime
@@ -35,6 +34,10 @@ ext {
     elasticSearchVersion  = "8.4.3"
 }
 
+task wrapper(type: Wrapper){
+    gradleVersion = '7.5.1'
+}
+
 dependencies {
 
     implementation project(":internal-module")
@@ -43,9 +46,15 @@ dependencies {
 
     implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-xml:2.14.0-rc2'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa:2.7.4'
-    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.14.0-rc2'
 
     implementation('org.springframework.boot:spring-boot-starter-web')
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
+
+}
+
+
+
+test {
+    useJUnitPlatform()
 }
