--- server/game-server/buildsrc/build.gradle
@@ -1,18 +1,0 @@
-plugins {
-    id 'groovy-gradle-plugin'
-}
-
-repositories {
-    gradlePluginPortal()
-}
-
-dependencies {
-    implementation 'gradle.plugin.com.github.spotbugs.snom:spotbugs-gradle-plugin:4.7.2'
-    testImplementation platform("org.spockframework:spock-bom:2.1-groovy-3.0")
-    testImplementation 'org.spockframework:spock-core'
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
-
