--- server/game-server/buildsrc/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name='jastelo-org-conventions'
