--- server/game-server/buildsrc/src/main/groovy/java-conventions.gradle
@@ -1,13 +1,0 @@
-plugins {
-    id 'java'
-}
-
-repositories {
-    mavenCentral()
-}
-
-
-// Enable deprecation messages when compiling Java code
-tasks.withType(JavaCompile).configureEach {
-    options.compilerArgs << "-Xlint:deprecation"
-}
