--- server/game-server/buildsrc/src/main/groovy/library-conventions.gradle
@@ -1,32 +1,0 @@
-plugins {
-    id 'java-library'
-    id 'maven-publish'
-    id 'java-conventions'
-}
-
-
-group = 'com.example'
-
-publishing {
-    publications {
-        library(MavenPublication) {
-            from components.java
-        }
-    }
-    repositories {
-        maven {
-            name "jasteloRepository"
-            url 'build/my-repo'
-        }
-    }
-}
-
-tasks.register('printProps') {
-    doLast {
-        println commandLineProjectProp
-        println gradlePropertiesProp
-        println systemProjectProp
-        println System.metaPropertyValues
-        println System.properties['system']
-    }
-}
