--- server/game-server/internal-module/build.gradle
@@ -1,21 +1,0 @@
-plugins {
-    id 'java-library'
-    id 'java-conventions'
-}
-
-group 'com.example'
-version '0.0.1-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-ext {
-    logbackVersion = '1.2.11'
-}
-dependencies {
-    implementation 'org.slf4j:jul-to-slf4j:2.0.3'
-    implementation "ch.qos.logback:logback-core:${logbackVersion}",
-            "ch.qos.logback:logback-classic:${logbackVersion}"
-    implementation 'org.json:json:20220924'
-}
