--- server/game-server/internal-module/src/main/java/com/example/model/GameDesignVars.java
@@ -1,27 +1,0 @@
-package com.example.model;
-
-public final class GameDesignVars {
-    public static final boolean USE_STUB_IN_PLACE_OF_SERVER = false;
-    public static final boolean ONLY_ONE_GAME = true;
-    public static final Integer GAME_LOBBY_ID = 1;
-    public static final Integer BAD_GAME_ID = 0;
-    public static final boolean START_WITH_SOUND_ENABLED = false;
-    public static final Integer MAX_GAME_FPS = 30;
-    public static final Integer MAX_PLAYERS_PER_GAME = 4;
-    public static final float LOBBY_SCREEN_TIME_BETWEEN_UPDATES = 3.0F;
-    public static final float LOBBY_SCREEN_UPDATE_RATE = 0.33333334F;
-    public static final float GAMEPLAY_MAX_SERVER_UPDATE_RATE = 5.0F;
-    public static final float GAMEPLAY_TIME_BETWEEN_UPDATES = 0.2F;
-    public static final int DEFAULT_PLAYER_ID = 0;
-    public static final String DEFAULT_PLAYER_NAME = "";
-    public static final int STARTING_LIVES = 3;
-    public static final int MAX_LIVES = 3;
-    public static final float MAX_PLAYER_SPEED = 2.0F;
-    public static final int STARTING_AMMO = 1;
-    public static final int MAX_AMMO = 1;
-    public static final float DAMAGE_FROM_HIT = 1.0F;
-    public static final float MAX_DISTANCE = 30.0F;
-
-    private GameDesignVars() {
-    }
-}
