--- server/game-server/settings.gradle
@@ -1,3 +1,0 @@
-rootProject.name = 'game-server'
-include 'internal-module'
-
