+++ server/game-server/src/main/java/com/example/game/server/exceptions/RestResponseEntityExceptionHandler.java
@@ -1,0 +1,38 @@
+package com.example.game.server.exceptions;
+
+import org.slf4j.*;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.method.annotation.*;
+
+@ControllerAdvice
+public class RestResponseEntityExceptionHandler extends ResponseEntityExceptionHandler {
+    private static final Logger logger;
+
+    static {
+        logger = LoggerFactory.getLogger(RestResponseEntityExceptionHandler.class);
+        logger.info("this.class/logger = {}", RestResponseEntityExceptionHandler.class.toString());
+    }
+//    @ExceptionHandler(value
+//            = { IllegalArgumentException.class, IllegalStateException.class })
+//    protected ResponseEntity<Object> handleConflict(
+//            RuntimeException ex, WebRequest request) {
+//        String bodyOfResponse = "This should be application specific";
+//        return handleExceptionInternal(ex, bodyOfResponse,
+//                new HttpHeaders(), HttpStatus.CONFLICT, request);
+//    }
+
+//
+//    @ExceptionHandler({ConstraintViolationException.class})
+//    public ResponseEntity<Object> handleBadRequest(final ConstraintViolationException ex, final WebRequest request) {
+//        final String bodyOfResponse = "This should be application specific";
+//        return handleExceptionInternal(ex, bodyOfResponse, new HttpHeaders(), HttpStatus.BAD_REQUEST, request);
+//    }
+//
+//
+//    @ExceptionHandler({NullPointerException.class, IllegalArgumentException.class, IllegalStateException.class})
+//        public ResponseEntity<Object> handleInternal(final RuntimeException ex, final WebRequest request) {
+//        logger.error("500 Status Code", ex);
+//        final String bodyOfResponse = "This should be application specific";
+//        return handleExceptionInternal(ex, bodyOfResponse, new HttpHeaders(), HttpStatus.INTERNAL_SERVER_ERROR, request);
+//    }
+}
