--- server/game-server/src/main/java/com/example/server/exceptions/MyResourceNotFoundException.java
@@ -1,26 +1,0 @@
-package com.example.server.exceptions;
-
-
-import org.springframework.http.*;
-import org.springframework.web.bind.annotation.*;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public final class MyResourceNotFoundException extends RuntimeException {
-
-    public MyResourceNotFoundException() {
-        super();
-    }
-
-    public MyResourceNotFoundException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-
-    public MyResourceNotFoundException(final String message) {
-        super(message);
-    }
-
-    public MyResourceNotFoundException(final Throwable cause) {
-        super(cause);
-    }
-
-}
