+++ server/game-server/src/main/resources/application.properties
@@ -1,7 +1,8 @@
+spring.application.name = ServerApplication
 spring.resources.add-mappings=true
 server.port=8080
 logging.level.root=info
-logging.level.com.example.game-server=trace
+logging.level=trace
 spring.profiles.active=dev
 logging.pattern.console=%d{dd-MM-yyyy HH:mm:ss.SSS} %magenta([%thread]) %highlight(%-5level) %logger.%M - %msg%n
 logging.path=logs
