--- server/game-server/src/main/resources/logback.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-    <property name="LOG_DIR" value="logs"/>
-
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>
-                %d{dd-MM-yyyy HH:mm:ss.SSS} %magenta([%thread]) %highlight(%-5level) %logger.%M - %msg%n
-            </pattern>
-        </encoder>
-    </appender>
-
-    <appender name="SAVE-TO-FILE" class="ch.qos.logback.core.FileAppender">
-        <file>${LOG_DIR}/log.log</file>
-        <encoder class="ch.qos.logback.classic.encoder.PatternLayoutEncoder">
-            <Pattern>
-                %d{dd-MM-yyyy HH:mm:ss.SSS} [%thread] %-5level %logger{36}.%M - %msg%n
-            </Pattern>
-        </encoder>
-    </appender>
-
-    <springProfile name="dev">
-        <root level="info">
-            <appender-ref ref="STDOUT"/>
-            <appender-ref ref="SAVE-TO-FILE"/>
-        </root>
-        <logger name="org.example.game-server" additivity="false" level="debug">
-            <appender-ref ref="STDOUT"/>
-            <appender-ref ref="SAVE-TO-FILE"/>
-        </logger>
-    </springProfile>
-
-    <springProfile name="prod">
-        <root level="info">
-            <appender-ref ref="SAVE-TO-FILE"/>
-        </root>
-        <logger name="org.example.internal-module" additivity="false" level="error">
-            <appender-ref ref="SAVE-TO-FILE"/>
-        </logger>
-    </springProfile>
-
-</configuration>
