+++ server/internal-module/build.gradle
@@ -1,0 +1,30 @@
+plugins {
+    id 'java-library'
+    id 'java-conventions'
+}
+
+group 'com.example'
+version '0.0.1-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+
+ext {
+    logbackVersion = '1.2.11'
+}
+
+
+dependencies {
+
+    implementation 'org.slf4j:jul-to-slf4j:2.0.3'
+    implementation "ch.qos.logback:logback-core:${logbackVersion}",
+            "ch.qos.logback:logback-classic:${logbackVersion}"
+
+    implementation 'jakarta.json.bind:jakarta.json.bind-api:3.0.0'
+    implementation 'jakarta.json:jakarta.json-api:2.1.1'
+
+    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.14.0-rc2'
+    implementation 'org.json:json:20220924'
+
+}
