+++ server/settings.gradle
@@ -1,0 +1,4 @@
+rootProject.name = 'server'
+include 'internal-module'
+include 'game-server'
+
